refactor: update to file names, added middleware dir
